requirements.txt
nltk
textblob
fastapi
uvicorn
jinja2
python-multipart
huggingface_hub
python-dotenv
transformers
torch==2.8.0 
torchvision==0.23.0
torchaudio==2.8.0
langdetect